+++ gosundpy/gosund.py
@@ -7,13 +7,19 @@ from .utils import cache_response
 class Gosund(object):
 
     def __init__(self, username, password, access_id, access_key, country=1,
-            endpoint='https://openapi.tuyaus.com'):
+            endpoint='https://openapi.tuyaus.com', status_cache_seconds=None):
         self.api = TuyaOpenAPI(endpoint, access_id, access_key,
                 auth_type=AuthType.CUSTOM)
         resp = self.api.connect(username, password, country)
         assert_response_success('connect to api', resp)
         self.manager = TuyaDeviceManager(self.api, TuyaOpenMQ(self.api))
         self._known_devices = {}
+
+        self._status_caching = False
+        if status_cache_seconds is not None:
+            self.get_device_statuses = cache_response(
+                    seconds=status_cache_seconds)(self.get_device_statuses)
+            self._status_caching = True
 
     def get_device(self, device_id):
         resp = self.manager.get_device_functions(device_id)
@@ -28,7 +34,6 @@ class Gosund(object):
                 f'unable to find status for device with id "{device_id}"')
         return status
 
-    @cache_response(seconds=60)
     def get_device_statuses(self):
         # limit 20 device_ids per api call
         resp = self.manager.get_device_list_status(self._known_devices)
@@ -44,7 +49,8 @@ class Gosund(object):
         self._clear_statuses_cache()
 
     def _clear_statuses_cache(self):
-        self.get_device_statuses.clear_cache()
+        if self._status_caching:
+            self.get_device_statuses.clear_cache()
 
     def send_commands(self, device_id, commands):
         resp = self.manager.send_commands(device_id, commands)