+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Unreleased
+### Bug Fixes
++ Make caching of `Gosund.get_device_statuses` optional via opt in and
+  configurable
 
 ## 0.6.1
 ### Bug Fixes